Major changes to prng and reorganized code
